chatgpt-plugin-config/requirements.txt
PyYAML
requests
flask
flask_cors